Refactor json usage and cleanup logic
