+++ node/rpc/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["ChainSafe Systems <info@chainsafe.io>"]
 edition = "2018"
 
 [dependencies]
-actor = { path = "../../vm/actor/",features = ["json"] }
+actor = { path = "../../vm/actor/" }
 async-std = { version = "1.6.3", features = ["attributes"] }
 serde = { version = "1.0.101", default-features = false, features = ["derive"] }
 serde_json = "1.0.48"
@@ -42,6 +42,7 @@ async-tungstenite = "0.9.1"
 async-log = "2.0.0"
 log ="0.4.8"
 flo_stream = "0.4.0"
+libp2p = { version = "0.24", default-features = false }
 
 [dev-dependencies]
 db = { path = "../db" }
