+++ vm/actor/Cargo.toml
@@ -6,14 +6,12 @@ edition = "2018"
 
 [features]
 v2 = ["ipld_hamt/v2"]
-json = []
 
 [dependencies]
 vm = { package = "forest_vm", path = "../../vm" }
 address = { package = "forest_address", path = "../address" }
 runtime = { path = "../runtime" }
 num-bigint = { path = "../../utils/bigint", package = "forest_bigint" }
-beacon = { path = "../../blockchain/beacon" }
 encoding = { package = "forest_encoding", path = "../../encoding" }
 num-traits = "0.2"
 num-derive = "0.3.0"
@@ -36,7 +34,6 @@ base64 = "0.13"
 log = "0.4.8"
 commcid = { path = "../../utils/commcid" }
 indexmap = { version = "1.3.2", features = ["serde-1"] }
-libp2p = { version = "0.24", default-features = false }
 
 [dev-dependencies]
 hex = "0.4.2"
